Add default parse method for Decodable collections
We added a default implementation for parse if the response type was
`Decodable`, but weren't able to do the same for collections of
decodable types, because of how Argo implemented the Array extension.
With thoughtbot/Argo#243, we can actually extend Request to
automatically parse ResponseTypes that are CollectionTypes whose
elements are themselves Decodable.