wip: Post notification if an error was encountered
This still needs tests

It's useful in some cases to get a notification when the server
encountered an error. For example, you could have a top-level object
listening for these network notifications that catches them, then either
logs them or displays them to the user.

This doesn't change the fact that the error is still returned as the
failure part of the result. That means that if something needs to happen
locally in response to the error, we don't prevent that.