Add JSON to server errors
Previously, we were throwing away the JSON that the server sent with
non-200 responses. This was short-sighted, since any good API should be
sending back JSON info about what went wrong in addition to the status
code.

This change tries to fix that by embedding the raw JSON response in the
returned error object under a special key. This JSON can then be used
for debugging, or for displaying to the user.