Add the ability to parse error objects
This adds a `transformError` function to `Request. This acts as a hook
to be able to change the error from the `NSError` used by the lower
level network clients and JSON parsing into a model more suitable for
your application.

By default, we are still going to return `NSError`s, but it's useful to
have a hook for being able to model the network error response if you
know what it looks like.

This maintains 100% backwards compatibility with the existing API.