Return data tasks when performing requests
Previously, we were throwing away the data tasks created by
NSURLSession. This made the API simpler, but the end result was that
there was no way to cancel a request once it started.

By returning the data task, we can now leave it up to the consumer on
how to handle canceling requests. If they want the old behavior, they
can simply continue to ignore the return result. But if they want, they
can now hold onto the task that is returned, and cancel it or do
whatever they need to do with it.