Create JSON errors under the Swish domain
Previously we were using the error constructor provided from Result when
converting `Decoded`, to `Result`. This resulted in vending `NSError`
instances with the Result domain. This felt wrong for a number of
reasons.

This change moves the construction of these errors to our internal
function, which means the error is now under our domain.