Wrap errors in SwishError instead of NSError
Previously, we were passing around instances of NSError internally. This
worked fine, but ultimately doesn't provide enough error info to be able
to be able to do any kind of meaningful transformations down the line.

By wrapping our errors in a SwishError enum, we can be more explicit
about where the errors are coming from, and what they represent.