Don't force completion blocks to be scheduled on the main queue
Different request performers are allowed to have different semantics
when it comes to the queue that results are delivered on. If a performer
already happens to deliver results on the main queue, `APIClient` is
injecting latency by scheduling an unnecessary async call. If a
performer delivers results on a background queue (i.e., `NetworkRequestPerformer`),
and there's further processing that would benefit from occurring in the
background (e.g., image processing or decoding a large object), callers
must dispatch back onto a background queue to avoid blocking the UI. It
also hampers composability of requests, because multiple chained
requests will incur a penalty from jumping back and forth from the main
queue to a background queue.

This breaking change removes the `dispatc_async`, replacing it with the
more abstract idea of a `Scheduler`, which is a block that schedules the
completion handler to be performed, either synchronously or
asynchronously. The `mainQueueScheduler` is the default, and an
`immediateScheduler` is provided to synchronously invoke the completion
block instead.