Don't lock transitive dependencies using submodules
When Carthage checks out Swish into a project, it also recursively
copies its submodules. This means that even if Carthage finds a more
recent version of Result or Argo, when Swish is built it will always
link against the version in the submodule.

If the app and Swish are linked against two different versions of the
same dependency, the app may crash on launch when looking up symbols.

This change now ignores Cartfile.resolved, and stops using submodules,
so that Swish can dynamically be linked against any compatible version
that the app uses, rather than a specific submodule revision.