Fix CustomNSError implementation on Swift 3.0.0
Since Swift 3.0.1, `CustomNSError` provides default implementations of
`errorCode` and `errorUserInfo`. Unfortunately, in order to support
Swift 3.0.0, we need to add a custom implementation. Further, it's not
possible to specify a patch version in `#if swift()`, so in order to use
the default implementation we have to target at least Swift 3.1 (even
though the default implementation is available in Swift 3.0.1).