Allow two requests to be performed concurrently
Adds a new method to the `Swish.Client` protocol, allowing two requests
to be performed concurrently and combining their results in a
thread-safe manner.

If either request fails with an error, the other is cancelled if it is
still in flight. The requests are performed on a global concurrent queue
of default priority.