Model form-urlencoded payloads as [URLQueryItem]
Modelling application/x-www-form-urlencoded payloads as `[String: String]`
is problematic, because it's perfectly valid to have query parameters
with no associated value. That is, `param=` and `param` are two distinct
query paramets: the first with an empty value, and the second with no
value.

This change replaces `[String: String]` with `[URLQueryItem]`, solving
the "no value" problem, and removing the problematic extra layer of URL
encoding. This simpler implementation doesn't make too many extra
guarantees, but still allows users to apply their own extra levels of
encoding/decoding when the manually construct or read instances of
`URLQueryItem`.