Assume Data and URLResponse are non-nil
The documentation for `URLSession.dataTask(with:completionHandler:)`
states that the `data` and `error` parameters to the completion handler
are mutually exclusive:

> If the request completes successfully, the `data` parameter of the
> completion handler block contains the resource data, and the `error`
> parameter is `nil`. If the request fails, the `data` parameter is
> `nil` and the `error` parameter contain information about the failure.
> If a response from the server is received, regardless of whether the
> request completes successfully or fails, the `response` parameter
> contains that information.

It also follows from this that if the `data` parameter is non-nil, the
`request` parameter will also be non-nil.

This change updates `HTTPResponse` to more accurately model this,
removing some unnecessary `nil` checks, and force-downcasts the
`URLResponse` to `HTTPURLResponse`.