Update Swish for Swift 5.0.
Using the latest Quick and Nimble was pretty much the only code change
required.

We no longer need a framework for Result (even though this is was a good
one). It's included in the standard library now, so the only thing we
needed was an extension to add value and error to keep the tests happy
(and anyone who was depending on those in their code).