Support Swift 5.0 (#123)
* Update Swish for Swift 5.0.

Using the latest Quick and Nimble was pretty much the only code change
required.

We no longer need a framework for Result (even though this is was a good
one). It's included in the standard library now, so the only thing we
needed was an extension to add value and error to keep the tests happy
(and anyone who was depending on those in their code).

* Adjust from feedback.

Needed to ignore a new file and I was incorrectly creating a new decoder
in my fix (not sure how this worked in my tests like this).

* Removing the need for Result.

Now that Swift 5 provides this, we don't need it.

* Eliminate Result extension.

This caused the compiler to crash initially because there's so much type
inferrence going on in the tests that it couldn't even report errors. I
think I got it worked out to keep the intent behind the test without
having the `value` and `error` fields.